src/discord_kg/extraction/llm_powered/requirements.txt
# LangGraph-based Discord Knowledge Graph Extraction Dependencies

# Core LLM APIs
openai>=1.0.0
anthropic>=0.8.0

# LangGraph workflow orchestration
langgraph>=0.0.40
langchain-core>=0.1.0

# Data processing
pandas>=1.3.0
numpy>=1.21.0
PyYAML>=6.0

# Text processing
regex>=2022.0.0

# Optional: Local embeddings for fallback
sentence-transformers>=2.2.0
scikit-learn>=1.1.0

# Cloud storage (optional)
boto3>=1.26.0

# Graph database (optional)  
neo4j>=5.0.0

# Additional dependencies for the new architecture
typing-extensions>=4.0.0  # For enhanced type hints
dataclasses>=0.6  # Python 3.6 compatibility for dataclasses